Reorganize code for OSV advisories
